pyproject.toml
[tool.poetry]
name = "iu-carrige"
version = "1.0.0"
description = ""
authors = ["halone <haloneprog@gmail.com>"]
readme = "README.md"
packages = [{include = "iu_carrige"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.110.0"
sqlalchemy = "^2.0.29"
#psycopg2-binary = "^2.9.9"
blinker = "^1.7.0"
loguru = "^0.7.2"
aioboto3 = "^12.3.0"
redis = "^5.0.3"
asgi-lifespan = "^2.1.0"
pytest = "^8.1.1"
httpx = "^0.27.0"
pytest-dependency = "^0.6.0"
pytest-asyncio = "^0.23.6"
iu-datamodels = {git = "https://github.com/Halone228/iu_datamodels"}
uvicorn = "^0.29.0"
python-multipart = "^0.0.9"
types-aioboto3-lite = {extras = ["essential"], version = "^12.3.0"}
aiocache = "^0.12.2"
psycopg = "^3.1.18"
faker = "^24.7.1"
polyfactory = "^2.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
test = "main:start_test"
dev = "main:start_dev"